:star: new(addos/centered): support centered decorator for vue
