+++ addons/centered/src/index.js
@@ -1,27 +1,9 @@
-import React from 'react';
+import ReactCentered from './react';
+import VueCentered from './vue';
 
-const style = {
-  position: 'fixed',
-  top: 0,
-  left: 0,
-  bottom: 0,
-  right: 0,
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  overflow: 'auto',
-};
-
-const innerStyle = {
-  margin: 'auto',
-};
-
-export default function(storyFn) {
-  return (
-    <div style={style}>
-      <div style={innerStyle}>
-        {storyFn()}
-      </div>
-    </div>
-  );
+let Centered = ReactCentered; // default for React
+if (window.STORYBOOK_ENV === 'vue') {
+  Centered = VueCentered;
 }
+
+export default Centered;
